{TCP,UART}CtrlIntfModule.cpp: fix overflow bug
There was a overflow bug in the internal buffer handling. If the
received command was larger than '_max_cmd_len - 1' [*] the thread would
try to write on invalid memory. Now, in that case, we simply ignore the
exceeding chars.

[*] The buffer size was increased by 1 so that the maximum command
length is given by '_max_cmd_len' (and not '_max_cmd_len - 1', which was
accounting for the string terminator).