TCPFwUpdateModule: add module
Add module that provides mechanisms for updating the bootloader or the
application remotely. It's designed to work with the newboot [1]
bootloader.

The application must define on its linker script the symbols
'__FWUpdateFlash_start', '__FWUpdateFlash_end' and '__FWInfo_addr'
(listed in newboot.h) and they must match the definitions on the newboot
linker script. This is inherited from openMMC [2] and there may be a
better way to share symbols.

[1] https://gitlab.cnpem.br/GIE/Projects/newboot
[2] https://github.com/lnls-dig/openMMC